Adds volume trace processing attributes
Adds reactive_compression_time, reactive_offset_points, nonreactive_offset_points,
reactive_end_time, and nonreactive_end_time as attributes of the Condition. Also
adds corresponding compression_time, offset_points, and output_end_time attributes
to the Experiment. Getting/setting the Condition attributes operates on the
attributes of the Experiment class instance representing the reactive_case or
nonreactive_case, depending on the attribute. This is so there is a sensible
interface from the Condition instance, but the actual values are stored with
the appropriate information.