Loading the YAML file now sets attributes and is optional
In all cases, the value of the attribute takes precedence for the work in
create_volume_trace. However, for convenience, values can still be read
from volume-trace.yaml, and these values are used to set the appropriate
attributes. The preferred format of the YAML file is now to use the attribute
names explicitly. However, old format files can be loaded using load_old_yaml.