Load CTI file in Condition constructor
The CTI file for Cantera should be specified as an argument to Condition.
If it is not specified, the user is prompted for the file name.
The file is then read into a string variable, which is preferred as the
initializer for Solutions. The cti_source is used to test for thermo
warnings, and then the warnings are suppressed for the remainder of the
session.

Fixes #3, #11