Reduce influence of DAQ startup effects
Sometimes the DAQ will record an erroneous voltage for the first sample,
so set the first sample equal to the mean of the first 200 samples. This
is important because the initial pressure depends on the mean of the
first few samples being at 0V (and the first point in particular should
be at 0V), so we need to correct this.

Work on #6.