Set reactive_case and nonreactive_case
Since these are no longer being set by the respective _file attributes,
need to set them where they're used. If the file name of any existing
case is different from the value specified in the _file attributes,
replace it with the one from the _file name so that the _case can be
changed.